refactor: component PageContent
