+++ src/page/TodoPage/TodoPage.module.css
@@ -6,14 +6,3 @@
 .aside {
   background-color: var(--todo-bg);
 }
-
-.content {
-  padding: 40px;
-  min-width: 480px;
-  width: 100%;
-  max-width: 720px;
-  justify-self: center;
-  display: flex;
-  flex-direction: column;
-  gap: 40px;
-}
