+++ src/page/TodoPage/components/PageContent/Description/index.ts
@@ -1,0 +1,2 @@
+import Description from './Description';
+export default Description;
