+++ src/page/TodoPage/components/PageContent/PageContent.props.ts
@@ -1,0 +1,4 @@
+export interface PageContentProps {
+  title: string;
+  description: string;
+}
