+++ src/page/TodoPage/components/PageContent/Title/index.ts
@@ -1,0 +1,2 @@
+import Title from './Title';
+export default Title;
