+++ src/page/TodoPage/index.ts
@@ -1,0 +1,2 @@
+import TodoPage from './TodoPage';
+export default TodoPage;
